refactor(bump): split into versioning/engine modules; add complexity & size gates (#207, #208)
#207 — Code structure:
- Extract pure version math to commands/bump_versioning.py (219 LOC) and the
  bump-my-version glue to commands/bump_engine.py (242 LOC). bump.py drops from
  1040 -> 698. Moved names are re-imported into bump.py so external imports and
  monkeypatch-by-path keep working; only do_bump/get_configuration patches (whose
  call sites moved into the engine) were repointed (11 sites in test_bump_command).
- Gate module size: tests/test_module_size.py (every commands/*.py <= 800 LOC).
- Gate complexity: enable ruff C901 (max-complexity=10). Refactored the two
  violators with no behavior change — bump_command (12) via _resolve_language +
  _finalize_bump, and _execute_rollback (11) via _delete_rollback_tags. PLR0913 is
  intentionally NOT enabled (it fights typer command signatures). C901 is ignored
  for tests/ (elaborate mock setups), consistent with the existing per-file-ignores.

#208 — Robustness:
- bump_engine.py is now the single module that touches bump-my-version internals
  (typed via a BumpConfig alias).
- tests/test_bumpversion_contract.py asserts the upstream Config.pre_commit_hooks
  field still exists and is list-typed, so an upstream change fails loudly in CI
  rather than at release time.

439 passed, 100% coverage; ty (strict), ruff (incl. C901), interrogate all green.

Closes #207
Closes #208
