+++ tests/test_bump_command.py
@@ -473,7 +473,7 @@ def test_bump_with_allow_dirty_flag(bump_project, monkeypatch):
 
         return real_get_config(*args, **kwargs)
 
-    monkeypatch.setattr("rhiza_tools.commands.bump.get_configuration", mock_get_config)
+    monkeypatch.setattr("rhiza_tools.commands.bump_engine.get_configuration", mock_get_config)
 
     # Test with allow_dirty=True
     bump_command(BumpOptions(version="patch", allow_dirty=True))
@@ -492,7 +492,7 @@ def test_bump_with_commit_flag(bump_project, monkeypatch):
 
         return real_get_config(*args, **kwargs)
 
-    monkeypatch.setattr("rhiza_tools.commands.bump.get_configuration", mock_get_config)
+    monkeypatch.setattr("rhiza_tools.commands.bump_engine.get_configuration", mock_get_config)
 
     # Test with commit=True
     bump_command(BumpOptions(version="patch", commit=True))
@@ -532,7 +532,7 @@ def test_bump_folds_changelog_hook_into_commit(bump_project, monkeypatch):
         if kwargs.get("dry_run") is False:
             captured["config"] = kwargs.get("config")
 
-    monkeypatch.setattr("rhiza_tools.commands.bump.do_bump", mock_do_bump)
+    monkeypatch.setattr("rhiza_tools.commands.bump_engine.do_bump", mock_do_bump)
 
     bump_command(BumpOptions(version="patch", commit=True))
 
@@ -549,7 +549,7 @@ def test_bump_without_cliff_config_adds_no_changelog_hook(bump_project, monkeypa
         if kwargs.get("dry_run") is False:
             captured["config"] = kwargs.get("config")
 
-    monkeypatch.setattr("rhiza_tools.commands.bump.do_bump", mock_do_bump)
+    monkeypatch.setattr("rhiza_tools.commands.bump_engine.do_bump", mock_do_bump)
 
     bump_command(BumpOptions(version="patch", commit=True))
 
@@ -563,7 +563,7 @@ def test_bump_configuration_load_failure(bump_project, monkeypatch):
     def mock_get_config(*args, **kwargs):
         raise Exception("Configuration load error")  # noqa: TRY002, TRY003
 
-    monkeypatch.setattr("rhiza_tools.commands.bump.get_configuration", mock_get_config)
+    monkeypatch.setattr("rhiza_tools.commands.bump_engine.get_configuration", mock_get_config)
 
     # Should fail with exit code 1
     with pytest.raises(typer.Exit) as excinfo:
@@ -577,7 +577,7 @@ def test_bump_operation_failure(bump_project, monkeypatch):
     def mock_do_bump(*args, **kwargs):
         raise Exception("Bump operation failed")  # noqa: TRY002, TRY003
 
-    monkeypatch.setattr("rhiza_tools.commands.bump.do_bump", mock_do_bump)
+    monkeypatch.setattr("rhiza_tools.commands.bump_engine.do_bump", mock_do_bump)
 
     # Should fail with exit code 1
     with pytest.raises(typer.Exit) as excinfo:
@@ -908,7 +908,7 @@ class TestPreflightValidation:
             # Should never reach the real call
             raise AssertionError("Real bump should not be called after preflight failure")  # noqa: TRY003
 
-        with patch("rhiza_tools.commands.bump.do_bump", side_effect=mock_do_bump):
+        with patch("rhiza_tools.commands.bump_engine.do_bump", side_effect=mock_do_bump):
             with pytest.raises(typer.Exit) as excinfo:
                 bump_command(BumpOptions(version="patch"))
             assert excinfo.value.exit_code == 1
@@ -930,7 +930,7 @@ class TestPreflightValidation:
 
             return real_do_bump(*args, **kwargs)
 
-        with patch("rhiza_tools.commands.bump.do_bump", side_effect=tracking_do_bump):
+        with patch("rhiza_tools.commands.bump_engine.do_bump", side_effect=tracking_do_bump):
             bump_command(BumpOptions(version="patch"))
 
         # Should have two calls: first dry-run (preflight), then real
@@ -951,7 +951,7 @@ class TestPreflightValidation:
 
             return real_do_bump(*args, **kwargs)
 
-        with patch("rhiza_tools.commands.bump.do_bump", side_effect=tracking_do_bump):
+        with patch("rhiza_tools.commands.bump_engine.do_bump", side_effect=tracking_do_bump):
             bump_command(BumpOptions(version="patch", dry_run=True))
 
         # Should have only one call (the actual dry-run), no preflight
@@ -1446,7 +1446,7 @@ class TestExecuteBump:
         mock_config_path = MagicMock()
 
         with (
-            patch("rhiza_tools.commands.bump.do_bump", side_effect=Exception("bump failed")),
+            patch("rhiza_tools.commands.bump_engine.do_bump", side_effect=Exception("bump failed")),
             pytest.raises(typer.Exit),
         ):
             _execute_bump("1.0.1", mock_config, mock_config_path, dry_run=False)
@@ -1459,7 +1459,7 @@ class TestExecuteBump:
         mock_config_path = MagicMock()
 
         with (
-            patch("rhiza_tools.commands.bump.do_bump", side_effect=Exception("bump failed")),
+            patch("rhiza_tools.commands.bump_engine.do_bump", side_effect=Exception("bump failed")),
             pytest.raises(typer.Exit),
         ):
             _execute_bump("1.0.1", mock_config, mock_config_path, dry_run=True)
